Check constraint made more complex
